Use CommonJS so we don't need Babel
