+++ package.json
@@ -8,8 +8,8 @@
     "url": "https://github.com/noflo/noflo-assembly.git"
   },
   "scripts": {
-    "pretest": "./node_modules/.bin/eslint index.es6.js example/components/*.js spec/*.js && ./node_modules/.bin/babel index.es6.js --out-file index.js",
-    "test": "./node_modules/.bin/mocha --compilers js:babel-core/register -R spec spec/*.js"
+    "pretest": "./node_modules/.bin/eslint index.es6.js example/components/*.js spec/*.js",
+    "test": "./node_modules/.bin/mocha -R spec spec/*.js"
   },
   "keywords": [
     "noflo"
@@ -24,9 +24,6 @@
     "noflo": "^1.0.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.23.1",
-    "babel-preset-es2015": "^6.22.0",
     "chai": "^4.1.2",
     "eslint": "^4.12.0",
     "eslint-config-airbnb-base": "^12.1.0",
